requirements.txt
torch==2.2.2
langchain==0.3.27
langchain-community==0.3.29
langchain-openai==0.3.33
chromadb==1.1.0
sentence-transformers==2.2.2
streamlit==1.50.0
pypdf==6.1.0
python-dotenv==1.1.1
safetensors==0.6.2
schedule==1.2.2